Consolidate https functions into 1 express app
This lets functions use path params, and encapsulates all https routing
for the app.

It also makes it easier to stand the webapp server up on its own.
This reduces its dependency on cloud functions, and allows it to be run
in other environments.

The downside is that this makes it harder to scale the resource allocation
for different endpoints. E.g. with one function per endpoint ones that
require more ram/cpu can be given it without causing other endpoints to
cost more. This project isn't at the point where that matters. If a high
frequency endpoint needed to be optimized for cost it would always be
broken out.